[commit message skipped]

+++ functions/image-upload-handler/main.go
@@ -1,8 +1,11 @@
 package main
 
 import (
+	"bytes"
 	"context"
 	"fmt"
+	"log"
+	"net/http"
 	"time"
 
 	"github.com/aws/aws-lambda-go/events"
@@ -13,14 +16,46 @@ import (
 	"github.com/aws/aws-sdk-go/aws/session"
 	"github.com/aws/aws-sdk-go/service/s3"
 
+	"github.com/kelseyhightower/envconfig"
 	"github.com/rwcarlsen/goexif/exif"
 )
 
-// TargetBucketName is the bucket to which the source photos will be copied.
-const TargetBucketName string = "js-photos"
+// EnvVars are user-provided variables that influence the Lambda function's
+// execution.
+type EnvVars struct {
+	// TargetBucketName is the bucket to which the source photos will be copied.
+	TargetBucketName       string `required:"true" split_words:"true"`
+	WillDeleteSourceObject bool   `default:"true" split_words:"true"`
+	CallbackWebhookURL     string `required:"false" split_words:"true"`
+}
+
+func parseEnvVars() (EnvVars, error) {
+	var envVars EnvVars
+	err := envconfig.Process("SIO", &envVars)
+	if err != nil {
+		return envVars, err
+	}
+	return envVars, nil
+}
+
+func invokeCallbackWebhook(callbackWebhookURL string) {
+	response, err := http.Post(callbackWebhookURL, "", bytes.NewBuffer([]byte("")))
+	if err != nil {
+		log.Println("Warning: a callback webhook was defined, but an error " +
+			"occurred trying to invoke it.")
+	}
+	defer response.Body.Close()
+}
 
 // HandleRequest processes the incoming S3 PutObject event.
 func HandleRequest(ctx context.Context, event events.S3Event) (string, error) {
+	// Parse the environment variables.
+	envVars, err := parseEnvVars()
+	if err != nil {
+		return fmt.Sprintf("Error: %s", err), err
+	}
+	fmt.Printf("%s\n%s\n", envVars.TargetBucketName, envVars.CallbackWebhookURL)
+
 	// Instantiate the S3 Clicnet.
 	s3Client := s3.New(session.New())
 	// Alias the S3 record for convenient referencing.
@@ -29,6 +64,9 @@ func HandleRequest(ctx context.Context, event events.S3Event) (string, error) {
 	// Extract the bucket name and object key from the lambda event...
 	bucketName := s3Record.Bucket.Name
 	objectKey := s3Record.Object.Key
+	objectPath := "/" + bucketName + "/" + objectKey
+
+	fmt.Printf("%s\t%s\n", bucketName, objectKey)
 	// And use these values to construct a `GetObjectInput` so we can retrieve
 	// the actual image from S3.
 	getObjInput := &s3.GetObjectInput{
@@ -54,11 +92,11 @@ func HandleRequest(ctx context.Context, event events.S3Event) (string, error) {
 	response := "Date created: " + datetime.Format(time.RFC3339)
 
 	dateString := datetime.Format("2006-01-02")
-	// isoString := datetime.Format(time.RFC3339)
 
+	targetBucketName := envVars.TargetBucketName
 	copyObjInput := &s3.CopyObjectInput{
-		Bucket:     aws.String(TargetBucketName),
-		CopySource: aws.String(fmt.Sprintf("/%s/%s", bucketName, objectKey)),
+		Bucket:     aws.String(targetBucketName),
+		CopySource: aws.String(objectPath),
 		Key: aws.String(
 			fmt.Sprintf("/%s/%s", dateString, objectKey),
 		),
@@ -81,19 +119,23 @@ func HandleRequest(ctx context.Context, event events.S3Event) (string, error) {
 		return fmt.Sprintf(copyResult.GoString()), err
 	}
 
-	deleteObjectInput := &s3.DeleteObjectInput{
-		Bucket: aws.String(bucketName),
-		Key:    aws.String(objectKey),
+	if envVars.WillDeleteSourceObject {
+		deleteObjectInput := &s3.DeleteObjectInput{
+			Bucket: aws.String(bucketName),
+			Key:    aws.String(objectKey),
+		}
+		deleteObjResult, err := s3Client.DeleteObject(deleteObjectInput)
+		fmt.Println(deleteObjResult)
+		if err != nil {
+			fmt.Println("Error deleting original object.")
+			return fmt.Sprintf("Error: "), err
+		}
 	}
 
-	deleteObjResult, err := s3Client.DeleteObject(deleteObjectInput)
-	fmt.Println(deleteObjResult)
-	if err != nil {
-		fmt.Println("Error deleting original object.")
-		return fmt.Sprintf("Error: "), err
+	webhookURL := envVars.CallbackWebhookURL
+	if envVars.CallbackWebhookURL != "" {
+		invokeCallbackWebhook(webhookURL)
 	}
-
-	fmt.Println(result)
 
 	return fmt.Sprintf(response), nil
 }
